Remove unnecessary assignments
There's no need to do these assignments. Assignment in Ruby happens to
be an expression (so the value of `a = b` is `b`), so the function still
works, but it's cleaner if you just return the result of the call to
`map`. Same deal in the else branch: `blob = blob` is unnecessary.